move config.js location
